final before defense

+++ libft/ft_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: nihamdan <nihamdan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 09:44:44 by nihamdan          #+#    #+#             */
-/*   Updated: 2023/05/04 14:23:27 by nihamdan         ###   ########.fr       */
+/*   Updated: 2023/05/18 17:45:03 by nihamdan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-t_list	*ft_lstnew(int content)
+t_list	*ft_lstnew(long content)
 {
 	t_list	*dst;
