+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: nihamdan <nihamdan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 09:48:32 by nihamdan          #+#    #+#             */
-/*   Updated: 2023/05/04 12:18:31 by nihamdan         ###   ########.fr       */
+/*   Updated: 2023/05/18 17:45:40 by nihamdan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
 typedef struct s_list
 {
 	size_t			index;
-	int				content;
+	long			content;
 	struct s_list	*next;
 }	t_list;
 
@@ -32,7 +32,7 @@ int		ft_toupper(int c);
 int		ft_tolower(int c);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 int		ft_lstsize(t_list *lst);
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
@@ -66,7 +66,7 @@ void	ft_lstiter(t_list *lst, void (*f)(int));
 size_t	ft_strlen(const char *s);
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize);
-t_list	*ft_lstnew(int content);
+t_list	*ft_lstnew(long content);
 t_list	*ft_lstlast(t_list	*lst);
 
 #endif
